refactor(ui): move navigation primitives into shared/ui/navigation
