+++ src/shared/ui/index.ts
@@ -26,11 +26,11 @@ export { IconSubmitButton } from "./forms/icon-submit-button";
 export { Loading, Spinner } from "./feedback/loading";
 export { ProgressBar } from "./feedback/progress-bar";
 export { StatBar } from "./stat-bar";
-export { Stepper } from "./stepper";
-export { TabNav } from "./tab-nav";
-export type { TabItem } from "./tab-nav";
-export { UnderlineTabs } from "./underline-tabs";
-export type { UnderlineTabItem } from "./underline-tabs";
+export { Stepper } from "./navigation/stepper";
+export { TabNav } from "./navigation/tab-nav";
+export type { TabItem } from "./navigation/tab-nav";
+export { UnderlineTabs } from "./navigation/underline-tabs";
+export type { UnderlineTabItem } from "./navigation/underline-tabs";
 export { Console, ConsoleTitleBar } from "./overlays/console";
 export { InfoBox } from "./feedback/info-box";
 export { RevealMark } from "./reveal-mark";
