+++ src/shared/ui/navigation/index.ts
@@ -1,0 +1,5 @@
+export { Stepper } from "./stepper";
+export { TabNav } from "./tab-nav";
+export type { TabItem } from "./tab-nav";
+export { UnderlineTabs } from "./underline-tabs";
+export type { UnderlineTabItem } from "./underline-tabs";
