+++ src/shared/ui/navigation/step-box.ts
