+++ src/shared/ui/navigation/stepper.tsx
